完善修改订单状态接口

+++ backend/db_course_design/Services/impl/HotelService.cs
@@ -45,7 +45,7 @@ namespace db_course_design.Services.impl
                     HotelId = (int)o.HotelId,
                     RoomType = o.RoomType,
                     RoomPrice = o.RoomPrice,
-                    RoomLeft = o.RoomLeft,
+                    RoomLeft = (int)o.RoomLeft,
                 }).ToListAsync();
             return RoomDetail;
         }
